Add more openOCD automation as a TCL script
This script is automatically loaded when launching 'openocd -f
openocd.cfg'.
It adds 'flash_bds', 'ramboot_bds' commands to the openOCD console. The
former is writing the current EC firmware inside the internal flash, the
latter is loading a RAM only firmware on the chip.
There are similar commands for the Link Proto-0 board :
'flash_link', 'ramboot_link'.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=none
TEST=type 'flash_bds' from openocd telnet console

Change-Id: Ic63281a99ee1d083986696675ed0a82da7c033ee