stm32l: fix length argument for i2c writes
Pass the actual msg_len instead of the maximum outbound message size.

BUG=none
TEST=tested on daisy using logic analyzer (notes below)

Beforehand, the output on a logic analyzer looked sane until the last
bits and stop condition, after which the bus appeared to go haywire
and the system would reset. I'm guessing that this is due to the EC
trying to drive SDA line while the host is holding SCL high.

Change-Id: I1afa520d858905a571540341fe02f374dcdf43c1
Signed-off-by: David Hendricks <dhendrix@chromium.org>