stm32l: clear and then set GPIO mode and pull-up/down settings
GPIO mode and pull-up/down registers do not all get initialized to
zero on reset. This patch ensures that all bits in the those registers
are set explicitly. An intermediate variable is used so that changes
are made atomically.

Note: output speed registers are also not all initialized to zero, but
we don't handle that in gpio_pre_init yet.

BUG=none
TEST=tested on newer daisy boards (which needed this patch to boot)

Change-Id: Ice2795197135dcee8f8484e4908dbfcf90fec2c9
Signed-off-by: David Hendricks <dhendrix@chromium.org>