Add header_size in struct jump_data.
The position of jump_tags was shifted every time a new field was added
to struct jump_data. This broke the sysjump hook badly.

To make this more scalable, add a header_size field in struct jump_data.
Then the new code can always prepend (or reduce if jump_data becomes smaller)
some spaces between jump_data and jump_tags.

BUG=chrome-os-partner:9447
TEST=EC upgrade from EC 517 (2231) to this version, and keyboard keeps working.
Note that EC 526 (2235) to this version is not working because we have no way
to identify that header version change.

Change-Id: If1b506c6f7d22e5affaaf8ada15990f60d2f957a