Remove the keyboard "cold reset" hack
This was needed because we could not do a clean
reset with CF9 and were relying on the keyboard
reset pulse to trigger a cold reset.

However it has the downside of causing a kernel
panic to be a cold reset and the kernel panic
information in memory is lost.

Now that we have a VR workaround this hack can
be disabled and the keyboard can instead issue
a warm reset.

BUG=chrome-os-partner:11036
TEST=manual

1) install this EC on a Link device and boot
2) log in and execute "echo panic > /proc/breakme"
3) after reboot look for kernel entry in /var/spool/crash

Change-Id: I1134618f0a202d37aaae69a1d857fa8853a7e52c
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/27722
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Randall Spangler <rspangler@chromium.org>